refactor inputs_handler/outputs_handler
