+++ genet/__init__.py
@@ -1,7 +1,7 @@
 from genet.core import Network  # noqa: F401
 from genet.schedule_elements import Schedule, Service, Route, Stop  # noqa: F401
 from genet.use.road_pricing import Toll  # noqa: F401
-from genet.inputs_handler.read import *  # noqa: F401,F403
+from genet.input.read import *  # noqa: F401,F403
 from genet.auxiliary_files import AuxiliaryFile  # noqa: F401
 from genet.max_stable_set import MaxStableSet  # noqa: F401
 from genet.exceptions import ScheduleElementGraphSchemaError, RouteInitialisationError, \
