+++ genet/input/osm_reader.py
@@ -4,10 +4,10 @@ import osmread
 from pyproj import Transformer
 from math import ceil
 
-import genet.inputs_handler.osmnx_customised as osmnx_customised
+import genet.input.osmnx_customised as osmnx_customised
 import genet.utils.parallel as parallel
 import genet.utils.spatial as spatial
-from genet.outputs_handler.matsim_xml_values import MATSIM_JOSM_DEFAULTS
+from genet.output.matsim_xml_values import MATSIM_JOSM_DEFAULTS
 
 
 class Config(object):
