+++ tests/test_inputs_handler_osmnx_customised.py
@@ -1,5 +1,5 @@
 from tests.fixtures import assert_semantically_equal, full_fat_default_config
-from genet.inputs_handler import osmnx_customised
+from genet.input import osmnx_customised
 
 
 def test_return_edges_handles_regular_non_oneway_paths(full_fat_default_config):
